requirements.txt
telebot
fpdf
Pil
image
arabic_reshaper
pyarabic.araby
tokenize
os
re